STT/pyproject.toml
[project]
name = "stt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "남도현", email = "2020097010@hanyang.ac.kr" }
]
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.1",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"
